refactor(habits): split habit calendar view into modular components
extract color helper logic into separate file and restructure
calendar view component for better maintainability and separation
of concerns
